Add clarification on how events are supported
So I have been working my way through adding stimulus to my application, and I stumbled upon the fact that it was actually somewhat unclear (to me, perhaps not others) how to support triggering other events.

Now I understand on the back side from what I can tell you're implementing the addEventListener API so that makes this easy, so I did figure out that you're simply short forming event->controller#method to the equivalent addEvenListener API. 

However, it did take me digging into the code a little to surmise this properly. I want to add a blurb to the docs that will make this clear and apparent.

The example is an example I found from an earlier issue, [issue 127](https://github.com/stimulusjs/stimulus/issues/127)

If its too verbose or whatever, let me know, but I think it would be great if the docs explicitly stated that this is how it works. Explicit vs implicit and all that jazz.