Document keeping controllers out of packs dir
Previously, the documentation for installation had an example implying a directory structure of `app/javascript/packs/controllers`. However, the documentation for rails/webpacker states that packs is a special directory that should only contain entrypoints to be included in HTML, ex. application.js.

This changes the installation example to imply a directory structure of `app/javascript/controllers`, and explicitly gives that structure as context to the example.