Account for HTML escaped ActionDescriptor Syntax
Some frameworks (e.g. Rails and [Rack][]) will HTML escape attributes.
This can result in typical action routing attributes to be transformed
and broken:

```html+erb
<button data-action="<%= an_escaping_method("input->controller#action")
%>"></button>

<%# => <button data-action="input-&gt;controller#action"></button>
```

This commit expands the Action Descriptor syntax to account for
escaping.

In the case of Rack, there is a small collection of HTML escaped
characters:

* `&` escapes into `&amp;`
* `<` escapes into `&lt;`
* `>` escapes into `&gt;`
* `'` escapes into `&#x27;`
* `"` escapes into `&quot;`
* `/` escapes into `&#x2F;`

Of those characters, the `>` is the only one currently part of the
Stimulus syntax.

[Rack]: https://www.rubydoc.info/github/rack/rack/Rack/Utils#ESCAPE_HTML-constant