return true for `has[name]Value` if value has custom default value
If you define a custom default value in your controller the `has[name]Value` method should return `true`. This is useful when you have behavior like this in your controller and want to differentiate between a custom default value and the regular default value:

  import { Controller } from 'stimulus'

  export default class extends Controller {
    static values = { refreshInterval: 500 }

    connect() {
      if (this.hasRefreshIntervalValue) {
        this.startRefreshing()
      }
    }
  }

So with this change the above code will be executed because you have a custom default value set.