Retain backtrace for TypeErrors in value change callback
This code was originally added to provide better error messaging
for TypeErrors when assigning values.

Unfortunately throwing a new error object causes us to lose
the context (including stacktrace) of the original error,
making it harder to locate the source.

Changing the message and rethrowing the original error
retain that information.