Allow for default method for controller actions
This allows for actions to omit the method. When the method is omitted
stimulus will use `_` as the method name. This will make it simpiler to
use single method controllers.

For example if we have a controller that disables an element before we
would write something like this:-

```html
  <button data-action="click->disable#disable">Disable on click</button>
```

```js
class DisableController extends Controller {
  disable({target}) {
    target.disabled = true
  }
}
```

After we can shorten the `data-action` attribute like this:-

```html
  <button data-action="click->disable">Disable on click</button>
```

```js
class DisableController extends Controller {
  _({target}) {
    target.disabled = true
  }
}
```

This makes it easier for developers to write the controller as they can
give the controller a clear name and not have to construct a name for
the method when its not needed.