Accept a rename object for outlets
Currently outlets are accept via an array of string `[a,b,c]`. This commit adds the ability to also accept a rename object withing this array that will determine how will the outlet be accessed.

For example

```
static outlets = ['alpha', {'outlet--with--very--long--name': short}]
```
now instead of `this.outletWithVeryLongNameOutlet` you can use `this.shortOutlet`